+++ src/main/java/com/prgrms/airbnb/domain/room/entity/SortTypeForHost.java
@@ -1,6 +1,5 @@
-package com.prgrms.airbnb.domain.room.dto;
+package com.prgrms.airbnb.domain.room.entity;
 
-import com.prgrms.airbnb.domain.room.entity.Room;
 import com.prgrms.airbnb.domain.room.repository.RoomRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -16,8 +15,7 @@ public enum SortTypeForHost {
     public Page<Room> findByHost(Long hostId, Pageable pageable, RoomRepository roomRepository) {
       return roomRepository.findByHostIdOrderByRating(hostId, pageable);
     }
-  }
-  ;
+  };
 
   abstract public Page<Room> findByHost(Long hostId, Pageable pageable, RoomRepository roomRepository);
 }
